Move history event metadata to its own module
